refactor(view): 패키지 이동
